Restructure packages. Renamed table.
