+++ mycore-base/src/main/java/org/mycore/backend/jpa/objectinfo/MCRObjectInfoEntityQueryResolver.java
@@ -16,7 +16,7 @@
  *  along with MyCoRe.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.mycore.backend.jpa;
+package org.mycore.backend.jpa.objectinfo;
 
 import java.sql.Date;
 import java.util.ArrayList;
@@ -25,6 +25,7 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import org.mycore.backend.jpa.MCREntityManagerProvider;
 import org.mycore.datamodel.classifications2.MCRCategLinkReference_;
 import org.mycore.datamodel.classifications2.MCRCategoryID;
 import org.mycore.datamodel.classifications2.MCRCategoryID_;
@@ -33,6 +34,7 @@ import org.mycore.datamodel.classifications2.impl.MCRCategoryImpl_;
 import org.mycore.datamodel.classifications2.impl.MCRCategoryLinkImpl;
 import org.mycore.datamodel.classifications2.impl.MCRCategoryLinkImpl_;
 import org.mycore.datamodel.common.MCRObjectIDDate;
+import org.mycore.datamodel.common.MCRObjectInfo;
 import org.mycore.datamodel.common.MCRObjectQuery;
 import org.mycore.datamodel.common.MCRObjectQueryResolver;
 import org.mycore.datamodel.ifs2.MCRObjectIDDateImpl;
