+++ mycore-base/src/main/resources/config/mycore.properties
@@ -114,7 +114,7 @@ MCR.Access.AddObjectDefaultRule=true
 # MCR.EventHandler.MCRObject.017.Class=org.mycore.datamodel.common.MCRServiceFlagEventHandler
 # Define XML store handler
   MCR.EventHandler.MCRObject.020.Class=org.mycore.datamodel.common.MCRXMLMetadataEventHandler
-  MCR.EventHandler.MCRObject.020A.Class=org.mycore.backend.jpa.MCRObjectEntityEventHandler
+  MCR.EventHandler.MCRObject.020A.Class=org.mycore.backend.jpa.objectinfo.MCRObjectEntityEventHandler
 # Define Static Content handler
   MCR.EventHandler.MCRObject.025.Class=org.mycore.services.staticcontent.MCRStaticContentEventHandler
 # Define Link Table handler
@@ -286,7 +286,7 @@ MCR.Access.Facts.Condition.category=org.mycore.access.facts.condition.fact.MCRCa
 
   MCR.Metadata.EnrichedDerivateLinkIDFactory.Class=org.mycore.datamodel.metadata.MCRDefaultEnrichedDerivateLinkIDFactory
 
-  MCR.Object.QueryResolver.Class=org.mycore.backend.jpa.MCRObjectInfoEntityQueryResolver
+  MCR.Object.QueryResolver.Class=org.mycore.backend.jpa.objectinfo.MCRObjectInfoEntityQueryResolver
 
 ##############################################################################
 # The configuration for write protection of website
